Issue #33: Add JQPL to Solr query translator Modified position field.

+++ MedSavantShared/src/main/java/org/ut/biolab/medsavant/shared/model/solr/SearcheableVariant.java
@@ -62,8 +62,8 @@ public class SearcheableVariant {
     }
 
     @Field("pos")
-    public void setPosition(Integer position) {
-        variantRecord.setPosition(Long.valueOf(position));
+    public void setPosition(Float position) {
+        variantRecord.setPosition(Long.valueOf(position.longValue()));
     }
 
     @Field("id")
